Implemented UrlShortener and InProcessDistributedCounter

+++ src/ShortUrl.Domain/Entities/ShortLink.cs
@@ -1,11 +1,11 @@
 ﻿namespace ShortUrl.Domain.Entities;
 
 public class ShortLink {
-    public long Id { get; }
+    public int Id { get; }
     public string OriginalUrl { get; }
     public string Token { get; }
 
-    public ShortLink(long id, string originalUrl, string token)
+    public ShortLink(int id, string originalUrl, string token)
     {
         Id = id;
         OriginalUrl = originalUrl;
